+++ CHANGELOG.md
@@ -7,11 +7,35 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+### Fixed
+
+- Unnecessary error logs when running a script that uses `env` in its shebang.
+- VSCode extension: running Python script with debugger fails because it tries to connect to the debugger port remotely.
+
 ### Changed
 
+- improved mirrord cli help message.
+
+## 3.30.0
+
+### Added
+
+- mirrord-layer: Added `port_mapping` under `incoming` configuration to allow mapping local ports to custom
+  remote port, for example you can listen on port 9999 locally and it will steal/mirror
+  the remote 80 port if `port_mapping: [[9999, 80]]`. See [#1129](https://github.com/metalbear-co/mirrord/issues/1129)
+
+### Fixed
+
+- Fix issue when two (or more) containerd sockets exist and we use the wrong one. Fixes [#1133](https://github.com/metalbear-co/mirrord/issues/1133).
+- Invalid toml in environment variables configuration examples.
+
+### Changed
+
+- Use container's runtime env instead of reading it from `/proc/{container_root_pid}/environ` as some processes (such as nginx) wipe it. Fixes [#1135](https://github.com/metalbear-co/mirrord/issues/1135)
 - Removed the prefix "test" from all test names - [#1065](https://github.com/metalbear-co/mirrord/issues/1065).
 - Created symbolic link from the vscode directory to the `LICENSE` and `CHANGELOG.md` files so that mirrord developers
   don't need to copy them there before building the app.
+- mirrord-layer: `socket` hook will now block ipv6 requests and will return EAFNOSUPPORT. See [#1121](https://github.com/metalbear-co/mirrord/issues/1121).
 
 ## 3.29.0
 
@@ -25,7 +49,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - Unpause the target container before exiting if the agent exits early on an error and the container is paused -
    [#1111](https://github.com/metalbear-co/mirrord/issues/1111).
 - intellij-plugin: fix issue where execution hangs when running using Gradle. Fixes [#1120](https://github.com/metalbear-co/mirrord/issues/1120).
-- intellij-plugin: fix issue where mirrord doesn't load into gradle, was found when fixing [#1120].
+- intellij-plugin: fix issue where mirrord doesn't load into gradle, was found when fixing [#1120](https://github.com/metalbear-co/mirrord/issues/1120).
 - mirrord-agent: reintroduce `-o lo` back to iptable rules to prevent issue where outinging messags could be intersepted by mirrord as incoming ones.
 - mirrord-layer: binding same port on different IPs leads to a crash due to `ListenAlreadyExists` error.
   This is now ignored with a `info` message since we can't know if the IP/Port was already bound